Always add a symlink to `panoptes.log`  for the `all` handler.

+++ pocs/utils/logger.py
@@ -249,14 +249,17 @@ def get_root_logger(profile='panoptes', log_config=None):
         # Create a symlink to the log file with just the name of the script and the handler
         # (level), as this makes it easier to find the latest file.
         # Use a relative path, so that if we move PANLOG the paths aren't broken.
-        log_symlink = os.path.join(log_dir, '{}-{}.log'.format(invoked_script, handler))
-        log_symlink_target = os.path.relpath(full_log_fname, start=log_dir)
-        try:
-            os.unlink(log_symlink)
-        except FileNotFoundError:  # pragma: no cover
-            pass
-        finally:
-            os.symlink(log_symlink_target, log_symlink)
+        for f in ['{}-{}.log'.format(invoked_script, handler), 'panoptes.log']:
+            if f == 'panoptes.log' and handler != 'all':
+                continue
+            log_symlink = os.path.join(log_dir, f)
+            log_symlink_target = os.path.relpath(full_log_fname, start=log_dir)
+            try:
+                os.unlink(log_symlink)
+            except FileNotFoundError:  # pragma: no cover
+                pass
+            finally:
+                os.symlink(log_symlink_target, log_symlink)
 
     # Configure the logger
     logging.config.dictConfig(log_config)
